Build out PATCH/UPDATE function
Added onUpdateItem, which will be activated the the "Update" button
is clicked. This will take in form data, and update the data object.

Added updateItem to API calls, which submits the AJAX PATCH request.

Tested using a curl script, and it was successful. Since the back-end
is not yet storing the updated title, we can't see the change on
the front-end.