Started building out update handlebar feature
Added a data-id to each instance of an update button
Created a placeholder function, called onSelectUpdate, which is
attached to the handlebars glyphicon-pencil button.
Used jQuery to retrieve and console log that id

Next steps are to build out a SHOW function so that we can bring
back data about a single upload and populate the update view.